+++ pyftpdlib/handlers.py
@@ -876,7 +876,7 @@ class DTPHandler(AsyncChat):
             # Close file object before responding successfully to client
             if self.file_obj is not None and not self.file_obj.closed:
                 self.file_obj.close()
-                
+
             if self._resp:
                 self.cmd_channel.respond(self._resp[0], logfun=self._resp[1])
 
@@ -3419,6 +3419,8 @@ if SSL is not None:
         certfile = None
         keyfile = None
         ssl_protocol = SSL.SSLv23_METHOD
+        # client certificate configurable attributes
+        client_certfile = None
         # - SSLv2 is easily broken and is considered harmful and dangerous
         # - SSLv3 has several problems and is now dangerous
         # - Disable compression to prevent CRIME attacks for OpenSSL 1.0+
@@ -3453,9 +3455,26 @@ if SSL is not None:
             self._pbsz = False
             self._prot = False
             self.ssl_context = self.get_ssl_context()
+            if self.client_certfile is not None:
+                from OpenSSL.SSL import VERIFY_CLIENT_ONCE
+                from OpenSSL.SSL import VERIFY_FAIL_IF_NO_PEER_CERT
+                from OpenSSL.SSL import VERIFY_PEER
+                self.ssl_context.set_verify(VERIFY_PEER |
+                                            VERIFY_FAIL_IF_NO_PEER_CERT |
+                                            VERIFY_CLIENT_ONCE,
+                                            self.verify_certs_callback)
 
         def __repr__(self):
             return FTPHandler.__repr__(self)
+
+        # Cannot be @classmethod, need instance to log
+        def verify_certs_callback(self, connection, x509,
+                                  errnum, errdepth, ok):
+            if not ok:
+                self.log("Bad client certificate detected.")
+            else:
+                self.log("Client certificate is valid.")
+            return ok
 
         @classmethod
         def get_ssl_context(cls):
@@ -3471,6 +3490,12 @@ if SSL is not None:
                 if not cls.keyfile:
                     cls.keyfile = cls.certfile
                 cls.ssl_context.use_privatekey_file(cls.keyfile)
+                if cls.client_certfile is not None:
+                    from OpenSSL.SSL import OP_NO_TICKET
+                    from OpenSSL.SSL import SESS_CACHE_OFF
+                    cls.ssl_context.load_verify_locations(cls.client_certfile)
+                    cls.ssl_context.set_session_cache_mode(SESS_CACHE_OFF)
+                    cls.ssl_options = cls.ssl_options | OP_NO_TICKET
                 if cls.ssl_options:
                     cls.ssl_context.set_options(cls.ssl_options)
             return cls.ssl_context